Move get_test_data into its own submodule
