+++ tests/test_hindcasting.py
@@ -5,8 +5,7 @@ from fiona.io import ZipMemoryFile
 
 from ravenpy.models import GR4JCN
 from ravenpy.utilities import forecasting
-
-from .common import get_test_data
+from ravenpy.utilities.testdata import get_test_data
 
 """
 Test to perform a hindcast using Caspar data on THREDDS.
